+++ bom/runtime/pom.xml
@@ -84,7 +84,7 @@
         <commons-codec.version>1.14</commons-codec.version>
         <classmate.version>1.3.4</classmate.version>
         <hibernate-orm.version>5.4.17.Final</hibernate-orm.version>
-	<hibernate-rx.version>1.0.0.Alpha5</hibernate-rx.version>
+        <hibernate-reactive.version>1.0.0.Alpha5</hibernate-reactive.version>
         <hibernate-validator.version>6.1.5.Final</hibernate-validator.version>
         <hibernate-search.version>6.0.0.Beta7</hibernate-search.version>
         <narayana.version>5.10.5.Final</narayana.version>
@@ -441,7 +441,7 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-hibernate-rx</artifactId>
+                <artifactId>quarkus-hibernate-reactive</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -2432,9 +2432,9 @@
                 <version>${hibernate-orm.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.hibernate.rx</groupId>
+                <groupId>org.hibernate.reactive</groupId>
                 <artifactId>hibernate-rx-core</artifactId>
-                <version>${hibernate-rx.version}</version>
+                <version>${hibernate-reactive.version}</version>
             </dependency>
             <dependency>
                 <groupId>org.hibernate</groupId>
