+++ extensions/hibernate-reactive/runtime/src/main/java/io/quarkus/hibernate/rx/runtime/RxPersistenceProviderSetup.java
