+++ extensions/hibernate-reactive/runtime/src/main/resources/META-INF/quarkus-extension.yaml
