--- extensions/hibernate-rx/runtime/pom.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>quarkus-hibernate-rx-parent</artifactId>
-        <groupId>io.quarkus</groupId>
-        <version>999-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>quarkus-hibernate-rx</artifactId>
-    <name>Quarkus - Hibernate RX - Runtime</name>
-    <description>TODO</description>
-    <dependencies>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-hibernate-orm</artifactId>
-        </dependency>
-        <dependency>
-        	<groupId>org.hibernate.rx</groupId>
-        	<artifactId>hibernate-rx-core</artifactId>
-        	<exclusions>
-        		<exclusion>
-        			<groupId>org.jboss.spec.javax.transaction</groupId>
-        			<artifactId>jboss-transaction-api_1.2_spec</artifactId>
-        		</exclusion>
-        	</exclusions>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-bootstrap-maven-plugin</artifactId>
-                <configuration>
-                    <excludedArtifacts>
-                        <excludedArtifact>javax.persistence:javax.persistence-api</excludedArtifact>
-                        <excludedArtifact>javax.persistence:persistence-api</excludedArtifact>
-                    </excludedArtifacts>
-                </configuration>
-            </plugin>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <annotationProcessorPaths>
-                        <path>
-                            <groupId>io.quarkus</groupId>
-                            <artifactId>quarkus-extension-processor</artifactId>
-                            <version>${project.version}</version>
-                        </path>
-                    </annotationProcessorPaths>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
