+++ extensions/pom.xml
@@ -84,7 +84,7 @@
         <module>agroal</module>
         <module>jdbc</module>
         <module>hibernate-orm</module>
-        <module>hibernate-rx</module>
+        <module>hibernate-reactive</module>
         <module>hibernate-validator</module>
         <module>panache</module>
         <module>hibernate-search-elasticsearch</module>
