+++ integration-tests/hibernate-reactive/src/test/java/io/quarkus/it/hibernate/rx/postgresql/HibernateReactiveInGraalIT.java
@@ -1,0 +1,7 @@
+package io.quarkus.it.hibernate.rx.postgresql;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+@NativeImageTest
+public class HibernateReactiveInGraalIT extends HibernateReactiveTest {
+}
