--- integration-tests/hibernate-rx/src/test/java/io/quarkus/it/jpa/postgresql/JPAFunctionalityInGraalITCase.java
@@ -1,7 +1,0 @@
-package io.quarkus.it.jpa.postgresql;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class JPAFunctionalityInGraalITCase extends JPAFunctionalityTest {
-}
