+++ integration-tests/pom.xml
@@ -44,7 +44,7 @@
         <module>hibernate-orm-panache</module>
         <module>hibernate-orm-rest-data-panache</module>
         <module>hibernate-orm-panache-kotlin</module>
-        <module>hibernate-rx</module>
+        <module>hibernate-reactive</module>
         <module>hibernate-search-elasticsearch</module>
         <module>hibernate-tenancy</module>
         <module>vertx-http</module>
